Move semantic events under controller
Place data-only semantic event types under internal/controller/event to
make their controller ownership explicit.

Update the runtime refactor plan to reflect the completed boundary
work and deferred registry and actor grouping decisions.
